support fetching chart data for user insights
