+++ propelauth_py/api/user.py
@@ -1,5 +1,4 @@
 from typing import Any, Optional, Dict, Union
-from propelauth_py.types.reports import UserReportType, UserReportRecord, UserReport
 from propelauth_py.user import _to_org_member_info
 import requests
 import httpx
@@ -835,126 +834,6 @@ async def _fetch_user_mfa_methods_async(
     mfa_setup = _deserialize_mfa_setup(json_response.get("mfa_setup"))
 
     return FetchUserMfaMethodsResponse(mfa_setup=mfa_setup)
-
-
-def _fetch_user_report(
-        auth_hostname,
-        integration_api_key,
-        report_key: UserReportType,
-        report_interval: Optional[str],
-        page_size: Optional[int],
-        page_number: Optional[int],
-    ) -> UserReport:
-    url = BASE_ENDPOINT_URL + "/user_report/" + report_key.value
-    params = {
-        "page_size": page_size,
-        "page_number": page_number,
-        "report_interval": report_interval,
-    }
-    response = requests.get(
-        url,
-        params=_format_params(params),
-        auth=_ApiKeyAuth(integration_api_key),
-        headers=_auth_hostname_header(auth_hostname),
-    )
-
-    if response.status_code == 401:
-        raise ValueError("integration_api_key is incorrect")
-    elif response.status_code == 429:
-        raise RateLimitedException(response.text)
-    elif response.status_code == 400:
-        raise BadRequestException(response.json())
-    elif not response.ok:
-        raise RuntimeError("Unknown error when fetching user reports")
-
-    json_response = response.json()
-    
-    user_report_records = [
-        UserReportRecord(
-            id=report_record.get("id"),
-            report_id=report_record.get("report_id"),
-            user_id=report_record.get("user_id"),
-            email=report_record.get("email"),
-            username=report_record.get("username"),
-            first_name=report_record.get("first_name"),
-            last_name=report_record.get("last_name"),
-            last_active_at=report_record.get("last_active_at"),
-            user_created_at=report_record.get("user_created_at"),
-            org_data=report_record.get("org_data"),
-            extra_properties=report_record.get("extra_properties"),
-        )
-        for report_record in json_response.get("user_reports")
-    ]
-    
-    return UserReport(
-        user_reports=user_report_records,
-        current_page=json_response.get("current_page"),
-        page_size=json_response.get("page_size"),
-        has_more_results=json_response.get("has_more_results"),
-        total_count=json_response.get("total_count"),
-        report_time=json_response.get("report_time"),
-    )
-
-async def _fetch_user_report_async(
-    httpx_client: httpx.AsyncClient,
-    auth_hostname,
-    integration_api_key,
-    report_key: UserReportType,
-    report_interval: Optional[str],
-    page_size: Optional[int],
-    page_number: Optional[int],
-) -> UserReport:
-    url = BASE_ENDPOINT_URL + "/user_report/" + report_key.value
-    params = {
-        "page_size": page_size,
-        "page_number": page_number,
-        "report_interval": report_interval,
-    }
-    response = await httpx_client.get(
-        url,
-        params=_format_params(params),
-        auth=_ApiKeyAuth(integration_api_key),
-        headers=_auth_hostname_header(auth_hostname),
-    )
-
-    if response.status_code == 401:
-        raise ValueError("integration_api_key is incorrect")
-    elif response.status_code == 429:
-        raise RateLimitedException(response.text)
-    elif response.status_code == 400:
-        raise BadRequestException(response.json())
-    elif not response.ok:
-        raise RuntimeError("Unknown error when fetching user reports")
-
-    response.raise_for_status()
-    json_response = response.json()
-    
-    user_report_records = [
-        UserReportRecord(
-            id=report_record.get("id"),
-            report_id=report_record.get("report_id"),
-            user_id=report_record.get("user_id"),
-            email=report_record.get("email"),
-            username=report_record.get("username"),
-            first_name=report_record.get("first_name"),
-            last_name=report_record.get("last_name"),
-            last_active_at=report_record.get("last_active_at"),
-            user_created_at=report_record.get("user_created_at"),
-            org_data=report_record.get("org_data"),
-            extra_properties=report_record.get("extra_properties"),
-        )
-        for report_record in json_response.get("user_reports")
-    ]
-    
-    return UserReport(
-        user_reports=user_report_records,
-        current_page=json_response.get("current_page"),
-        page_size=json_response.get("page_size"),
-        has_more_results=json_response.get("has_more_results"),
-        total_count=json_response.get("total_count"),
-        report_time=json_response.get("report_time"),
-    )
-
 
 ####################
 #       POST     #
